refactor: rename translation to translate
