+++ site/src/utils/translate.ts
